some cleaning done
